requirements.txt
streamlit>=1.31.0
langgraph>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
python-docx>=1.1.0
pdfplumber==0.11.7
